requirements.txt
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
bing-image-downloader==1.1.2
certifi==2025.1.31
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
h11==0.14.0
idna==3.10
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.4.2
jupyter-client==8.6.3
jupyter-core==5.7.2
kiwisolver==1.4.8
markupsafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prompt-toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.1
pyparsing==3.2.3
pysocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.3.0
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.30.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
typing-extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
